requirements.txt
aws-cdk.aws-codebuild~=1.48.0
aws-cdk.aws-codecommit~=1.48.0
aws-cdk.aws-codepipeline-actions~=1.48.0
aws-cdk.aws-codepipeline~=1.48.0
aws-cdk.core~=1.48.0
PyYAML~=5.3.1
typeguard~=2.9.1
